requirements.txt
discord.py
python-dotenv
requests
flask
google-cloud-documentai
firebase-admin